Complete TrendCoin separate component
