+++ src/Pages/Explore/index.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import { useDispatch } from "react-redux";
 import CoinData from "./CoinData/index";
-import TrendCoin from "./TrendCoin";
+import TrendCoin from "./TrendCoin/index";
 import { updatePageName } from "../../Redux/Actions/actionCreator";
 
 const Explore = () => {
